% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIEC.R
\name{createColors}
\alias{createColors}
\title{createColors}
\usage{
createColors(variable)
}
\arguments{
\item{variable}{character vector or factor variable.}
}
\value{
A named vector containing the color codes.
}
\description{
Produce a qualitative set of colors.
}
