% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordance.R
\name{extractStatistics}
\alias{extractStatistics}
\title{extractStatistics}
\usage{
extractStatistics(object, slot = "pValMat", colName = "rawP", type = "pvalue")
}
\arguments{
\item{object}{Output of the differential abundance tests on the comparisons
of Subset1 and Subset2. Must follow a specific structure list with subset,
comparison, and method levels. For each method, \code{pValMat},
\code{statInfo} matrices and method's \code{name} must be present (See
vignette for detailed information).}

\item{slot}{A character vector with 1 or number-of-methods-times repeats of
the slot names where to extract values for each method (default = "pValMat").}

\item{colName}{A character vector with 1 or number-of-methods-times repeats
of the column name of the slot where to extract values for each method
(default = "rawP").}

\item{type}{A character vector with 1 or number-of-methods-times repeats
of the value type of the column selected where to extract values for each
method. Two values are possible: "pvalue" or "logfc" (default = "pvalue").}
}
\value{
A list of \code{Subset1} and \code{Subset2} lists. Every subset
contains several comparisons. Each comparison contains the \code{colname}
column values, transformed according to \code{type} (not tranformed if
\code{type = "pvalue"}, \code{-abs(value)} if \code{type = "logfc"}), of the
\code{slot} result of DA analysis for each method.
}
\description{
Extract the list of p-values or logFold changes from the outputs of the
differential abundance detection methods to study the concordance within and
between methods.
}
\examples{
data("ps_plaque_16S")

set.seed(123)
phyloseq::sample_data(ps_plaque_16S)$HMP_BODY_SUBSITE <- factor(
    phyloseq::sample_data(ps_plaque_16S)$HMP_BODY_SUBSITE)
# At least N = 100 is suggested
split_list <- createSplits(object = ps_plaque_16S, varName =
    "HMP_BODY_SUBSITE", paired = "RSID", balanced = TRUE, N = 10)
# Run some methods
# lapply -> Subset1 and Subset2
Plaque_16S_splitsDA <- lapply(X = split_list, FUN = function(subset){
    # apply -> Comparison1, Comparison2, ..., Comparison10
    apply(X = subset, MARGIN = 1, FUN = function(splits) {
        # Splitting
        ps <- phyloseq::prune_samples(phyloseq::sample_names(ps_plaque_16S)
            \%in\% splits, ps_plaque_16S)
        # Keep only present taxa
        ps <- phyloseq::filter_taxa(ps, function(x) sum(x>0)>0, 1)
        # Adding scaling and normalization factors
        ps <- norm_edgeR(object = ps, method = "TMM")
        ps <- norm_CSS(object = ps, method = "median")
        # DA analysis
        returnList = list()
        returnList = within(returnList, {
            da.limma <- DA_limma(object = ps, design = ~ 1 +
                HMP_BODY_SUBSITE, coef = 2, norm = "TMM")
            da.limma.css <- DA_limma(object = ps, design = ~ 1 +
                HMP_BODY_SUBSITE, coef = 2, norm = "CSSmedian")
        })
        return(returnList)
    })
})

# Extract p-values
extracted_pvalues = extractStatistics(object = Plaque_16S_splitsDA, slot =
    "pValMat", colName = "rawP", type = "pvalue")
# Extract log fold changes
extracted_logfc = extractStatistics(object = Plaque_16S_splitsDA, slot =
    "statInfo", colName = "logFC", type = "logfc")
# Extract log fold changes for a method and p-values for the other
extracted_logfc_pvalues = extractStatistics(object = Plaque_16S_splitsDA,
    slot = c("statInfo", "pValMat"), colName = c("logFC", "rawP"), type =
    c("logfc","pvalue"))
}
\seealso{
[benchdamic::createConcordance()]
}
