% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{MDPlot}
\alias{MDPlot}
\title{MDPlot}
\usage{
MDPlot(data, difference = NULL, split = TRUE)
}
\arguments{
\item{data}{a `data.frame` object with a 'Model', 'Y', 'Y0', 'MD', and 'ZPD'
columns containing the model name, the observed values for the mean and the
zero proportion and the differences between observed and estimated values.}

\item{difference}{character vector, either 'MD' or 'ZPD' to plot the
differences between estimated and observed mean counts or the differences
between estimated zero probability and observed zero proportion.}

\item{split}{Display each model mean differences in different facets (default
= TRUE). If FALSE, points are not displayed for a more clear representation.}
}
\value{
a `ggplot` object.
}
\description{
A function to plot mean difference (MD) and zero probability difference (ZPD)
values between estimated and observed values.
}
\seealso{
\code{\link{fitModels}} and \code{\link{RMSE}} for the model
estimations and the RMSE computations respectively.
}
