% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{meanDifferences}
\alias{meanDifferences}
\title{meanDifferences}
\usage{
meanDifferences(estimated, observed)
}
\arguments{
\item{estimated}{a two column data.frame, output of `fitNB`, `fitZINB`,
`fitDM`, `fitZIG`, or `fitHURDLE` functions. More in general, a data frame
containing the continuity corrected logarithm for the average of the fitted
values for each row of a matrix of counts in the `Y` column, and the
estimated probability to observe a zero in the `Y0` column.}

\item{observed}{a two column data.frame, output of `prepareObserved`
function. More in general, a data frame containing
the continuity corrected logarithm for the average of the observed values for
each row of a matrix of counts in the `Y` column, and the estimated
proportion of zeroes in the `Y0` column.}
}
\value{
`data.frame` containing the differences between the estimated and the
observed continuity corrected logarithms of the average count values in the
`MD` columns, and between the estimated average probability to observe a zero
and the the observed zero rate in the `ZPD` column.
}
\description{
Compute the differences between the estimated and the observed continuity
corrected logarithms of the average count values (MD), and between the
estimated average probability to observe a zero and the the observed zero
rate (ZPD).
}
