% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbial_metabolism.R
\docType{data}
\name{microbial_metabolism}
\alias{microbial_metabolism}
\title{(Data) Microbial metabolism}
\format{
A \code{data.frame} object
}
\usage{
data(microbial_metabolism)
}
\description{
Aerobic, Anaerobic, or Facultative Anaerobic microbes by genus
(\href{https://github.com/waldronlab/nychanesmicrobiome}{NYC-HANES study}).
}
\keyword{datasets}
