% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordance.R
\name{plotConcordanceDendrogram}
\alias{plotConcordanceDendrogram}
\title{plotConcordanceDendrogram}
\usage{
plotConcordanceDendrogram(hc, direction = "v", cols)
}
\arguments{
\item{hc}{Hierarchical clustering results produced in
[benchdamic::plotConcordance()] function.}

\item{direction}{vertical (default \code{direction = "v"}) or horizontal
(\code{direction = "h"}).}

\item{cols}{A named vector containing the color hex codes.}
}
\value{
a \code{ggplot2} object
}
\description{
Plots the method's dendrogram of concordances.
}
\seealso{
[benchdamic::createConcordance()] and [bechdamic::plotConcordance()]
}
\keyword{internal}
