% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordance.R
\name{plotConcordanceHeatmap}
\alias{plotConcordanceHeatmap}
\title{plotConcordanceHeatmap}
\usage{
plotConcordanceHeatmap(c_df, threshold, cols)
}
\arguments{
\item{c_df}{A simplified concordance \code{data.frame} produced in
[benchdamic::plotConcordance()] function.}

\item{threshold}{The threshold for rank (x-axis upper limit if all methods
have a higher number of computed statistics).}

\item{cols}{A named vector containing the color hex codes.}
}
\value{
a \code{ggplot2} object
}
\description{
Plots the heatmap of concordances.
}
\seealso{
[benchdamic::createConcordance()] and [bechdamic::plotConcordance()]
}
\keyword{internal}
