% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIEC.R
\name{plotFPR}
\alias{plotFPR}
\title{plotFPR}
\usage{
plotFPR(df_FPR, cols = NULL)
}
\arguments{
\item{df_FPR}{5 columns per methods x comparisons rows data.frame produced by
the [bechdamic::createTIEC()] function, containing the FPR values.}

\item{cols}{named vector of colors.}
}
\value{
A ggplot object.
}
\description{
Draw the boxplots of the proportions of p-values lower than 0.01, 0.05, and
0.1 thresholds for each method.
}
