% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIEC.R
\name{plotKS}
\alias{plotKS}
\title{plotKS}
\usage{
plotKS(df_KS, cols = NULL)
}
\arguments{
\item{df_KS}{5 columns per methods x comparisons rows data.frame produced by
the [benchdamic::createTIEC] function.}

\item{cols}{named vector of colors.}
}
\value{
A ggplot object.
}
\description{
Draw the boxplots of the Kolmogorov-Smirnov test statistics for the p-value
distributions across the mock comparisons.
}
