% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOF.R
\name{plotMD}
\alias{plotMD}
\title{plotMD}
\usage{
plotMD(data, difference = NULL, split = TRUE)
}
\arguments{
\item{data}{a list, output of the \code{\link{fitModels}} function or a
`data.frame` object with \code{Model}, \code{Y}, \code{Y0}, \code{MD}, and
\code{ZPD} columns containing the model name, the observed values for the
mean and the zero proportion and the differences between observed and
estimated values.}

\item{difference}{character vector, either \code{MD} or \code{ZPD} to plot
the differences between estimated and observed mean counts or the differences
between estimated zero probability and observed zero proportion.}

\item{split}{Display each model mean differences in different facets
(default \code{split = TRUE}). If \code{FALSE}, points are not displayed for
more clear representation.}
}
\value{
a \code{ggplot} object.
}
\description{
A function to plot mean difference (MD) and zero probability difference (ZPD)
values between estimated and observed values.
}
\examples{
# Generate some random counts
counts = matrix(rnbinom(n = 60, size = 3, prob = 0.5), nrow = 10, ncol = 6)

# Estimate the counts assuming several distributions
GOF <- fitModels(
    counts = counts, models = c(
        "NB", "ZINB",
        "DM", "ZIG", "HURDLE"
    ), scale_ZIG = c("median", "default"), scale_HURDLE =
        c("median", "default")
)

# Plot the results
plotMD(data = GOF, difference = "MD", split = TRUE)
plotMD(data = GOF, difference = "ZPD", split = TRUE)
}
\seealso{
\code{\link{fitModels}} and \code{\link{RMSE}} for the model
estimations and the RMSE computations respectively. \code{\link{plotRMSE}}
for the graphical evaluation of the RMSE values.
}
