% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIEC.R
\name{plotQQ}
\alias{plotQQ}
\title{plotQQ}
\usage{
plotQQ(df_QQ, cols = NULL, zoom = c(0, 0.1))
}
\arguments{
\item{df_QQ}{Coordinates to draw the QQ-plot to compare the mean observed
p-value distribution across comparisons, with the theoretical uniform
distribution.}

\item{cols}{named vector of colors.}

\item{zoom}{2-dimesional vector containing the starting and the
final coordinates (default: \code{c(0, 0.1)})}
}
\value{
A ggplot object.
}
\description{
Draw the average QQ-plots across the mock comparisons.
}
