% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{plotRMSE}
\alias{plotRMSE}
\title{plotRMSE}
\usage{
plotRMSE(data, difference = NULL)
}
\arguments{
\item{data}{the `list` output of the \code{fitModels} function or a
`data.frame` object with a 'Model', 'Y', 'Y0', 'MD', and 'ZPD' columns
containing the model name, the observed values for the mean and the zero
proportion and the differences between observed and estimated values.}

\item{difference}{character vector, either 'MD' or 'ZPD' to plot the RMSE
values for the differences between estimated and observed mean counts or the
differences between estimated zero probability and observed zero proportion.}
}
\value{
a `ggplot` object.
}
\description{
A function to plot RMSE values computed for mean difference (MD) and zero
probability difference (ZPD) values between estimated and observed values.
}
\examples{
\donttest{
data(ps_stool_16S)

# Estimate the counts assuming several distributions
GOF_stool_16S <- fitModels(object = ps_stool_16S, models = c("NB","ZINB",
"DM","ZIG","HURDLE"), scale_ZIG = c("median","default"), scale_HURDLE =
c("median","default"))

# Plot the RMSE results
plotRMSE(data = GOF_stool_16S, difference = "MD")
plotRMSE(data = GOF_stool_16S, difference = "ZPD")
}
}
\seealso{
\code{\link{fitModels}} and \code{\link{RMSE}} for the model
estimations and the RMSE computations respectively. \code{\link{plotMD}} for
the graphical evaluation.
}
