% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{prepareObserved}
\alias{prepareObserved}
\title{prepareObserved}
\usage{
prepareObserved(counts, scale = NULL)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}

\item{scale}{If specified it refers to the character vector used in
`fitHURDLE` function. Either 'median' or 'default' to choose between the
median library size or one million as scaling factors for raw counts.}
}
\value{
A data frame containing the continuity corrected logarithm for the
raw count mean values for each taxon of the `counts` matrix in the `Y` column
and the observed zero rate in the `Y0` column. If `scale` is specified the
continuity corrected logarithm for the mean CPM (`scale = "default"`) or the
mean counts per median library size (`scale = "median"`) is computed instead.
}
\description{
Average count value continuity corrected logarithms and fraction of zeroes by
taxon.
}
\examples{
\dontrun{
data(ps_stool_16S)
counts = as(otu_table(ps_stool_16S), "matrix")

# Fit some models on the counts matrix
HURDLE <- fitHURDLE(counts, scale = "median")
NB <- fitNB(counts)

# Then log-transform the observed counts
# For HURDLE model remember to specify the "scale" parameter
observed_for_HURDLE <- prepareObserved(counts, scale = "median")
# For NB, ZINB, ZIG, and DM distributions  the "scale" parameter is not necessary
observed <- prepareObserved(counts)
}
}
\seealso{
\code{\link{meanDifferences}}
}
