% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plaque_16S.R
\docType{data}
\name{ps_plaque_16S}
\alias{ps_plaque_16S}
\title{(Data) 60 Gingival Plaque samples of 16S rRNA (HMP 2012)}
\format{
An object of class \code{phyloseq}
}
\usage{
data(ps_plaque_16S)
}
\description{
A demonstrative purpose dataset containing microbial abundances
for a total of 88 OTUs. The 60 Gingival Plaque paired samples belong to the
Human Microbiome Project. This particular subset contains 30 Supragingival
and 30 Subgingival Plaque samples from the \code{SEX = "Male"},
\code{RUN_CENTER = "WUCG"}, and \code{VISITNO = "1"} samples. It is possible
to obtain the same dataset after basic filters (remove taxa with zero counts)
and collapsing the counts to the genus level; \code{HMP16SData} Bioconductor
package was used to download the data.
}
\keyword{datasets}
