% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_stool_16S.R
\docType{data}
\name{ps_stool_16S}
\alias{ps_stool_16S}
\title{(Data) 33 Stool samples of 16S rRNA (HMP 2012)}
\format{
An object of class \code{phyloseq}
}
\usage{
data(ps_stool_16S)
}
\description{
A demonstrative purpose dataset containing microbial abundances
for a total of 71 OTUs. The 32 Stool samples belong to the Human Microbiome
Project. This particular subset contains the \code{SEX = "Male"},
\code{RUN_CENTER = "BI"}, and \code{VISITNO = "1"} samples. It is possible to
obtain the same dataset after basic filters (remove taxa with zero counts)
and collapsing the counts to the genus level; \code{HMP16Data} Bioconductor
package was used to download the data.
}
\keyword{datasets}
