% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(differences)
}
\arguments{
\item{differences}{a vector of differences.}
}
\value{
RMSE value
}
\description{
Computes the Root Mean Square Error (RMSE) from a vector of differences.
}
