% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{RMSEPlot}
\alias{RMSEPlot}
\title{RMSEPlot}
\usage{
RMSEPlot(data, difference = NULL)
}
\arguments{
\item{data}{a `data.frame` object with a 'Model', 'Y', 'Y0', 'MD', and 'ZPD'
columns containing the model name, the observed values for the mean and the
zero proportion and the differences between observed and estimated values.}

\item{difference}{character vector, either 'MD' or 'ZPD' to plot the RMSE
values for the differences between estimated and observed mean counts or the
differences between estimated zero probability and observed zero proportion.}
}
\value{
a `ggplot` object.
}
\description{
A function to plot RMSE values computed for mean difference (MD) and zero
probability difference (ZPD) values between estimated and observed values.
}
\seealso{
\code{\link{fitModels}} and \code{\link{RMSE}} for the model
estimations and the RMSE computations respectively.
}
