% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{runSplits}
\alias{runSplits}
\title{runSplits}
\usage{
runSplits(split_list, method_list, normalization_list, object)
}
\arguments{
\item{split_list}{A list of 2 \code{data.frame} objects: \code{Subset1} and
\code{Subset2} produced by the \code{\link{createSplits}} function.}

\item{method_list}{a list object containing the methods and their parameters.}

\item{normalization_list}{a list object containing the normalization method
names and their parameters produced by \code{\link{setNormalizations}}.}

\item{object}{a phyloseq object.}
}
\value{
A named list containing the results for each method.
}
\description{
Run the differential abundance detection methods on split datasets.
}
\examples{
data(ps_plaque_16S)

# Balanced design for independent samples
my_splits <- createSplits(
    object = ps_plaque_16S, varName =
    "HMP_BODY_SUBSITE", balanced = TRUE, N = 10 # N = 100 suggested
)

# Initialize some limma based methods
my_limma <- set_limma(design = ~ HMP_BODY_SUBSITE, coef = 2,
    norm = c("TMM", "CSSmedian"))

# Set the normalization methods according to the DA methods
my_norm <- setNormalizations(fun = c("norm_edgeR", "norm_CSS"),
    method = c("TMM", "median"))

# Run methods on split datasets
results <- runSplits(split_list = my_splits, method_list = my_limma,
    normalization_list = my_norm, object = ps_plaque_16S)
}
