% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{weights_ZINB}
\alias{weights_ZINB}
\title{weights_ZINB}
\usage{
weights_ZINB(
  object,
  design,
  K = 0,
  commondispersion = TRUE,
  zeroinflation = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{phyloseq object containing the counts and the sample data.}

\item{design}{character name of the metadata columns, formula, or design
matrix with rows corresponding to samples and columns to coefficients to be
estimated (the user needs to explicitly include the intercept in the design).}

\item{K}{integer. Number of latent factors.}

\item{commondispersion}{Whether or not a single dispersion for all features
is estimated (default TRUE).}

\item{zeroinflation}{Whether or not a ZINB model should be fitted. If FALSE,
a negative binomial model is fitted instead.}

\item{verbose}{Print helpful messages.}

\item{...}{Additional parameters to describe the model, see
\code{\link[zinbwave]{zinbModel}}.}
}
\value{
A matrix of weights.
}
\description{
Computes the observational weights of the counts under a zero-inflated
negative binomial (ZINB) model. For each count, the ZINB distribution is
parametrized by three parameters: the mean value and the dispersion of the
negative binomial distribution, and the probability of the zero component.
}
\examples{
data(ps_stool_16S)
# Calculate the ZINB weights
zinbweights <- weights_ZINB(object = ps_stool_16S, K = 0, design = "~ 1")
head(zinbweights)
}
\seealso{
\code{\link[zinbwave]{zinbFit}} for zero-inflated negative binomial
parameters' estimation and
\code{\link[zinbwave]{computeObservationalWeights}} for weights extraction.
}
