% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcordance.R
\name{areaCAT}
\alias{areaCAT}
\title{areaCAT}
\usage{
areaCAT(concordance, plotIt = FALSE)
}
\arguments{
\item{concordance}{A long format \code{data.frame} produced by
\link{createConcordance} function.}

\item{plotIt}{Plot the concordance (default \code{plotIt = FALSE}).}
}
\value{
A long format \code{data.frame} object with several columns:
\itemize{
    \item{\code{comparison}}{ which indicates the comparison number;}
    \item{\code{n_features }}{ which indicates the total number of taxa in
    the comparison dataset;}
    \item{\code{method1}}{ which contains the first method name;}
    \item{\code{method2}}{ which contains the first method name;}
    \item{\code{rank}}{;}
    \item{\code{concordance}}{ which is defined as the cardinality of the
    intersection of the top rank elements of each list, divided by rank, i.e.
    , \eqn{(L_{1:rank} \bigcap M_{1:rank})/(rank)}, where L and M represent
    the lists of the extracted statistics of method1 and method2
    respectively;}
    \item{\code{heightOver}}{ which is the distance between the bisector and
    the concordance value;}
    \item{\code{areaOver}}{ which is the cumulative sum of the
    \code{heightOver} value.}}
}
\description{
Compute the area between the bisector and the concordance curve.
}
\examples{
data("ps_plaque_16S")

set.seed(123)
phyloseq::sample_data(ps_plaque_16S)$HMP_BODY_SUBSITE <- factor(
    phyloseq::sample_data(ps_plaque_16S)$HMP_BODY_SUBSITE)
split_list <- createSplits(object = ps_plaque_16S,
                           varName = "HMP_BODY_SUBSITE",
                           paired = "RSID",
                           balanced = TRUE,
                           N = 10) # At least 100 is suggested

# Run some methods
# lapply -> Subset1 and Subset2
Plaque_16S_splitsDA <- lapply(X = split_list, FUN = function(subset){
    # apply -> Comparison1, Comparison2, ..., Comparison10
    apply(X = subset, MARGIN = 1, FUN = function(splits) {
        # Splitting
        ps <- phyloseq::prune_samples(phyloseq::sample_names(ps_plaque_16S)
            \%in\% splits, ps_plaque_16S)
        # Keep only present taxa
        ps <- phyloseq::filter_taxa(ps, function(x) sum(x>0)>0, 1)
        # Adding scaling and normalization factors
        ps <- norm_edgeR(object = ps, method = "TMM")
        ps <- norm_CSS(object = ps, method = "median")
        # DA analysis
        returnList = list()
        returnList = within(returnList, {
            da.limma <- DA_limma(object = ps, design = ~ 1 +
                HMP_BODY_SUBSITE, coef = 2, norm = "TMM")
            da.limma.css <- DA_limma(object = ps, design = ~ 1 +
                HMP_BODY_SUBSITE, coef = 2, norm = "CSSmedian")
        })
        return(returnList)
    })
})

# Concordance for p-values
concordance_pvalues = createConcordance(object = Plaque_16S_splitsDA, slot =
    "pValMat", colName = "rawP", type = "pvalue")

# Add area over the concordance curve
concordance_area <- areaCAT(concordance = concordance_pvalues)
}
\seealso{
\code{\link{createConcordance}} and \code{\link{plotConcordance}}
}
