% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIEC.R
\name{createTIEC}
\alias{createTIEC}
\title{createTIEC}
\usage{
createTIEC(object)
}
\arguments{
\item{object}{Output of the differential abundance tests on mock comparisons.
Must follow a specific structure with comparison, method, matrix of
p-values, and method's name (See vignette for detailed information).}
}
\value{
A \code{list} of \code{data.frame}s:
\itemize{
 \item{\code{df_pval}}{3 columns per number_of_features x methods x
 comparisons rows data.frame. The three columns are called Comparison, pval,
 and method;}
 \item{\code{df_FDR}}{5 columns per methods x comparisons rows data.frame.
 For each set of method and comparison, the proportion of false discoveries,
 considering 3 threshold (0.01, 0.05, 0.1) are reported;}
 \item{\code{df_QQ}}{contains the coordinates to draw the QQ-plot to compare
 the mean observed p-value distribution across comparisons, with the
 theoretical uniform distribution;}
 \item{\code{df_KS}}{5 columns and methods x comparisons rows data.frame. For
 each set of method and comparison, the Kolmogorov-Smirnov test statistics
 and p-values are reported in KS and KS_pval columns respectively.}
}
}
\description{
Extract the list of p-values from the outputs of the differential abundance
detection methods to compute several statistics to study the ability to
control the type I error.
}
\seealso{
[benchdamic::createMocks()]
}
