% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_corncob}
\alias{DA_corncob}
\title{DA_corncob}
\usage{
DA_corncob(
  object,
  pseudo_count = FALSE,
  formula,
  phi.formula,
  formula_null,
  phi.formula_null,
  test,
  boot = FALSE,
  coefficient = NULL,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSSmedian", "CSSdefault")
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{formula}{an object of class \code{formula} without the response: a symbolic description of the model to be fitted to the abundance}

\item{phi.formula}{an object of class \code{formula} without the response: a symbolic description of the model to be fitted to the dispersion}

\item{formula_null}{Formula for mean under null, without response}

\item{phi.formula_null}{Formula for overdispersion under null, without response}

\item{test}{Character. Hypothesis testing procedure to use. One of \code{"Wald"} or \code{"LRT"} (likelihood ratio test).}

\item{boot}{Boolean. Defaults to \code{FALSE}. Indicator of whether or not to use parametric bootstrap algorithm. (See \code{\link[corncob]{pbWald}} and \code{\link[corncob]{pbLRT}}).}

\item{coefficient}{The coefficient of interest as a single word formed by the
variable name and the non reference level. (e.g.: 'ConditionDisease' if the
reference level for the variable 'Condition' is 'control').}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}
}
\value{
A list object containing the matrix of p-values, the matrix of
summary statistics for each tag, and a suggested name of the final object
considering the parameters passed to the function.
}
\description{
Fast run for corncob differential abundance detection method.
}
\seealso{
\code{\link[corncob]{bbdml}} and
\code{\link[corncob]{differentialTest}} for differential abundance and
differential variance evaluation.
}
