% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_edgeR}
\alias{DA_edgeR}
\title{DA_edgeR}
\usage{
DA_edgeR(
  object,
  pseudo_count = FALSE,
  group = NULL,
  design = NULL,
  contrast = NULL,
  robust = FALSE,
  coef = 2,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSS"),
  weights
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{group}{vector or factor giving the experimental group/condition for each sample/library.}

\item{design}{numeric design matrix. Defaults to \code{model.matrix(~group)} if \code{group} is specified and otherwise to a single column of ones.}

\item{contrast}{numeric vector or matrix specifying one or more contrasts of the linear model coefficients to be tested equal to zero.}

\item{robust}{logical, should the estimation of \code{prior.df} be robustified against outliers?}

\item{coef}{integer or character index vector indicating which coefficients of the linear model are to be tested equal to zero. Ignored if \code{contrast} is not \code{NULL}.}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}

\item{weights}{optional numeric matrix giving observation weights}
}
\value{
A list object containing the matrix of p-values, the dispersion
estimates, the matrix of summary statistics for each tag, and a suggested
name of the final object considering the parameters passed to the function.
}
\description{
Fast run for edgeR differential abundance detection method.
}
\seealso{
[edgeR::DGEList()] for the edgeR DEG object creation,
[edgeR::estimateDisp()] and [edgeR::estimateGLMRobustDisp()] for dispersion
estimation, and [edgeR::glmQLFit()] and [edgeR::glmQLFTest()] for the
quasi-likelihood negative binomial model fit.
}
