% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_limma}
\alias{DA_limma}
\title{DA_limma}
\usage{
DA_limma(
  object,
  pseudo_count = FALSE,
  design = NULL,
  coef = 2,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSSmedian", "CSSdefault"),
  weights
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{design}{character name of the metadata columns, formula, or design
matrix with rows corresponding to samples and columns to coefficients to be
estimated.}

\item{coef}{column number or column name specifying which coefficient or contrast of the linear model is of interest. For \code{topTable}, can also be a vector of column subscripts, in which case the gene ranking is by F-statistic for that set of contrasts.}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}

\item{weights}{non-negative precision weights. Can be a numeric matrix of
individual weights of same size as the object expression matrix, or a numeric
vector of array weights with length equal to `ncol` of the expression matrix,
or a numeric vector of feature weights with length equal to `nrow `of the
expression matrix.}
}
\value{
A list object containing the matrix of p-values, the matrix of
summary statistics for each tag, and a suggested name of the final object
considering the parameters passed to the function.
}
\description{
Fast run for limma voom differential abundance detection method.
}
\seealso{
\code{\link[limma]{voom}} for the mean-variance relationship
estimation, \code{\link[limma]{lmFit}} for the linear model framework.
}
