% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_MAST}
\alias{DA_MAST}
\title{DA_MAST}
\usage{
DA_MAST(
  object,
  pseudo_count = FALSE,
  rescale = c("median", "default"),
  design,
  coefficient = NULL,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSSmedian", "CSSdefault")
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{rescale}{Rescale count data, per million if 'default', or per median
library size if 'median' ('median' is suggested for metagenomics data).}

\item{design}{The model for the count distribution. Can be the variable name,
or a character similar to "~ 1 + group", or a formula, or a `model.matrix`
object.}

\item{coefficient}{The coefficient of interest as a single word formed by the
variable name and the non reference level. (e.g.: 'ConditionDisease' if the
reference level for the variable 'Condition' is 'control')}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}
}
\value{
A list object containing the matrix of p-values, the matrix of
summary statistics for each tag, and a suggested name of the final object
considering the parameters passed to the function.
}
\description{
Fast run for MAST differential abundance detection method.
}
\seealso{
\code{\link[MAST]{zlm}} for the Truncated Gaussian Hurdle model
estimation.
}
