% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_metagenomeSeq}
\alias{DA_metagenomeSeq}
\title{DA_metagenomeSeq}
\usage{
DA_metagenomeSeq(
  object,
  pseudo_count = FALSE,
  design = NULL,
  coef = 2,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSSmedian", "CSSdefault")
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{design}{The model for the count distribution. Can be the variable name,
or a character similar to "~ 1 + group", or a formula, or a `model.matrix`
object.}

\item{coef}{Column number(s) or column name(s) specifying which coefficient
or contrast of the linear model to display.}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}
}
\value{
A list object containing the matrix of p-values, the matrix of
summary statistics for each tag, and a suggested name of the final object
considering the parameters passed to the function.
}
\description{
Fast run for the metagenomeSeq's differential abundance detection method.
}
\seealso{
\code{\link[metagenomeSeq]{fitZig}} for the Zero-Inflated Gaussian
regression model estimation and \code{\link[metagenomeSeq]{MRfulltable}}
for results extraction.
}
