% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{DA_Seurat}
\alias{DA_Seurat}
\title{DA_Seurat}
\usage{
DA_Seurat(
  object,
  pseudo_count = FALSE,
  test.use = "wilcox",
  group.by = NULL,
  reference.level = NULL,
  norm = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none",
    "ratio", "poscounts", "iterate", "TSS", "CSS")
)
}
\arguments{
\item{object}{phyloseq object.}

\item{pseudo_count}{Add 1 to all counts if TRUE (default = FALSE).}

\item{test.use}{Denotes which test to use. Available options are:
\itemize{
 \item{"wilcox"} : Identifies differentially expressed genes between two
 groups of cells using a Wilcoxon Rank Sum test (default)
 \item{"bimod"} : Likelihood-ratio test for single cell gene expression,
 (McDavid et al., Bioinformatics, 2013)
 \item{"roc"} : Identifies 'markers' of gene expression using ROC analysis.
 For each gene, evaluates (using AUC) a classifier built on that gene alone,
 to classify between two groups of cells. An AUC value of 1 means that
 expression values for this gene alone can perfectly classify the two
 groupings (i.e. Each of the cells in cells.1 exhibit a higher level than
 each of the cells in cells.2). An AUC value of 0 also means there is perfect
 classification, but in the other direction. A value of 0.5 implies that
 the gene has no predictive power to classify the two groups. Returns a
 'predictive power' (abs(AUC-0.5) * 2) ranked matrix of putative differentially
 expressed genes.
 \item{"t"} : Identify differentially expressed genes between two groups of
 cells using the Student's t-test.
 \item{"negbinom"} : Identifies differentially expressed genes between two
  groups of cells using a negative binomial generalized linear model.
  Use only for UMI-based datasets
 \item{"poisson"} : Identifies differentially expressed genes between two
  groups of cells using a poisson generalized linear model.
  Use only for UMI-based datasets
 \item{"LR"} : Uses a logistic regression framework to determine differentially
 expressed genes. Constructs a logistic regression model predicting group
 membership based on each feature individually and compares this to a null
 model with a likelihood ratio test.
 \item{"MAST"} : Identifies differentially expressed genes between two groups
 of cells using a hurdle model tailored to scRNA-seq data. Utilizes the MAST
 package to run the DE testing.
 \item{"DESeq2"} : Identifies differentially expressed genes between two groups
 of cells based on a model using DESeq2 which uses a negative binomial
 distribution (Love et al, Genome Biology, 2014).This test does not support
 pre-filtering of genes based on average difference (or percent detection rate)
 between cell groups. However, genes may be pre-filtered based on their
 minimum detection rate (min.pct) across both cell groups. To use this method,
 please install DESeq2, using the instructions at
 https://bioconductor.org/packages/release/bioc/html/DESeq2.html
}}

\item{group.by}{the name of a factor variable to group samples.}

\item{reference.level}{the level of the factor variable to take as reference
to compute differential abundance.}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis.}
}
\value{
A list object containing the matrix of p-values, the matrix of
summary statistics for each tag, and a suggested name of the final object
considering the parameters passed to the function.
}
\description{
Fast run for Seurat differential abundance detection method.
}
\seealso{
[Seurat::CreateSeuratObject] to create the Seurat object,
[Seurat::AddMetaData()] to add metadata information,
[Seurat::NormalizeData()] to compute the normalization for the counts,
[Seurat::FindVariableFeatures()] to estimate the mean-variance trend,
[Seurat::ScaleData()] to scale and center features in the dataset, and
[Seurat::FindMarkers()] to perform differential abundance analysis.
}
