% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitDM}
\alias{fitDM}
\title{fitDM}
\usage{
fitDM(counts)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}
}
\value{
A data frame containing the continuity corrected logarithms of the
average fitted values for each row of the `counts` matrix in the `Y` column,
and the estimated probability to observe a zero in the `Y0` column.
}
\description{
Fit a Dirichlet-Multinomial (DM) distribution for each taxon of the count
data. The model estimation procedure is performed by MGLM `MGLMreg` function
without assuming the presence of any group in the samples (design matrix
equal to a column of ones.)
}
\examples{
# Let's generate some random count data
nlibs <- 10
ntaxa <- 100
lambda = 2
counts <- matrix(rpois(ntaxa*nlibs, lambda),
                 nrow = ntaxa,
                 ncol = nlibs)
# DM model estimation
f_values <- fitDM(counts)
head(f_values)
}
