% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitHURDLE}
\alias{fitHURDLE}
\title{fitHURDLE}
\usage{
fitHURDLE(counts, scale = "default")
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}

\item{scale}{Character vector, either 'median' or 'default' to choose between
the median of the library size or one million to scale raw counts.}
}
\value{
A data frame containing the continuity corrected logarithms of the
average fitted values for each row of the `counts` matrix in the `Y` column,
and the estimated probability to observe a zero in the `Y0` column.
}
\description{
Fit a truncated gaussian hurdle model for each taxon of the count data. The
hurdle model estimation procedure is performed by MAST `zlm` function without
assuming the presence of any group in the samples (design matrix equal to a
column of ones.)
}
\examples{
# Let's generate some random count data
nlibs <- 10
ntaxa <- 100
lambda = 2
counts <- matrix(rpois(ntaxa*nlibs, lambda),
                 nrow = ntaxa,
                 ncol = nlibs)
# HURDLE model estimation
f_values <- fitHURDLE(counts,"median")
head(f_values)
}
