% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitModels}
\alias{fitModels}
\title{fitModels}
\usage{
fitModels(
  counts,
  models = c("NB", "ZINB", "DM", "ZIG", "HURDLE"),
  scale_ZIG = c("default", "median"),
  scale_HURDLE = c("default", "median")
)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}

\item{models}{character vector which assumes the values 'NB', 'ZINB', 'DM',
'ZIG', and 'HURDLE'.}

\item{scale_ZIG}{character vector, either 'median' or 'default' to choose
between the median of the library size or one thousand to scale normalization
factors for the zero-inflated gaussian model.}

\item{scale_HURDLE}{character vector, either 'median' or 'default' to choose
between the median of the library size or one million to scale raw counts for
the truncated gaussian hurdle model.}
}
\value{
list of `data.frame` objects for each model. The first two columns
contain the properly transformed observed values for mean and zero
proportion, while the third and the fourth columns contain the estimated
values for the mean and the zero rate resplectively.
}
\description{
A wrapper function that fits the specified models for each taxon of the count
data and computes the mean difference (MD) and zero probability difference
(ZPD) between estimated and observed values.
}
\seealso{
\code{\link{fitNB}}, \code{\link{fitZINB}}, \code{\link{fitDM}},
\code{\link{fitZIG}}, and \code{\link{fitHURDLE}} for the model estimations,
\code{\link{prepareObserved}} for raw counts preparation, and
\code{\link{meanDifferences}} for the Mean Difference (MD) and Zero
Probability Difference (ZPD) computations.
}
