% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitNB}
\alias{fitNB}
\title{fitNB}
\usage{
fitNB(counts)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}
}
\value{
A data frame containing the continuity corrected logarithms of the
average fitted values for each row of the `counts` matrix in the `Y` column,
and the estimated probability to observe a zero in the `Y0` column.
}
\description{
Fit a Negative Binomial (NB) distribution for each taxon of the count data.
The NB estimation procedure is performed by edgeR `glmFit` function, using
"TMM" normalized counts, tag-wise dispersion estimation, and not assuming the
presence of any group in the samples (design matrix equal to a column of
ones.)
}
\examples{
# Let's generate some random count data
nlibs <- 10
ntaxa <- 100
lambda = 2
counts <- matrix(rpois(ntaxa*nlibs, lambda),
                 nrow = ntaxa,
                 ncol = nlibs)
# NB estimation
f_values <- fitNB(counts)
head(f_values)
}
