% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitZIG}
\alias{fitZIG}
\title{fitZIG}
\usage{
fitZIG(counts, scale = "default")
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}

\item{scale}{Character vector, either 'median' or 'default' to choose between
the median of the library size or one thousand to scale normalization
factors.}
}
\value{
A data frame containing the continuity corrected logarithms of the
average fitted values for each row of the `counts` matrix in the `Y` column,
and the estimated probability to observe a zero in the `Y0` column.
}
\description{
Fit a Zero-Inflated Gaussian (ZIG) distribution for each taxon of the count
data. The model estimation procedure is performed by metagenomeSeq `fitZig`
function without assuming the presence of any group in the samples (design
matrix equal to a column of ones.)
}
\examples{
data(ps_stool_16S)
counts = as(phyloseq::otu_table(ps_stool_16S), "matrix")

# Fit model on the counts matrix
ZIG <- fitZIG(counts, scale = "median")
head(ZIG)
}
