% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{fitZINB}
\alias{fitZINB}
\title{fitZINB}
\usage{
fitZINB(counts)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}
}
\value{
A data frame containing the continuity corrected logarithms of the
average fitted values for each row of the `counts` matrix in the `Y` column,
and the estimated probability to observe a zero in the `Y0` column.
}
\description{
Fit a Zero-Inflated Negative Binomial (ZINB) distribution for each taxon of
the countdata. The ZINB estimation procedure is performed by zinbwave
`zinbFit` function with `commondispersion = FALSE`, regularization parameter
`epsilon = 1e10`, and not assuming the presence of any group in the samples
(design matrix equal to a column of ones.)
}
\examples{
\dontrun{
data(ps_stool_16S)
counts = as(otu_table(ps_stool_16S), "matrix")

# Fit model on the counts matrix
ZINB <- fitZINB(counts)
head(ZINB)
}
}
