% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_CSS}
\alias{norm_CSS}
\title{norm_CSS}
\usage{
norm_CSS(object, method = "default")
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}

\item{method}{Normalization scaling parameter (default = "1000"). If
"median", the median of the normalization factors is used as scaling (Paulson
et al. 2013).}
}
\value{
A new column containing the CSS scaling factors is added to the
phyloseq `sample_data` slot.
}
\description{
Calculate scaling factors from a phyloseq object to scale the raw library
sizes. Inherited from metagenomeSeq `calcNormFactors` function which performs
the Cumulative Sum Scaling normalization.
}
\examples{
data(ps_stool_16S)
# Calculate the scaling factors
ps_stool_16S <- norm_CSS(object = ps_stool_16S, method = "median")

# The phyloseq object now contains the scaling factors:
scaleFacts <- phyloseq::sample_data(ps_stool_16S)[, "NF.CSSmedian"]
head(scaleFacts)

# VERY IMPORTANT: to convert scaling factors to normalization factors
# multiply them by the library sizes and renormalize.
normFacts = scaleFacts * phyloseq::sample_sums(ps_stool_16S)
# Renormalize: multiply to 1
normFacts = normFacts/exp(mean(log(normFacts)))
}
\seealso{
\code{\link[metagenomeSeq]{calcNormFactors}} for details.
}
