% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_CSS}
\alias{norm_CSS}
\title{norm_CSS}
\usage{
norm_CSS(object, method = "default")
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}

\item{method}{Normalization scaling parameter (default = "1000"). If
"median", the median of the normalization factors is used as scaling (Paulson
et al. 2013).}
}
\value{
A new column containing the CSS normalization factors is added to the
phyloseq `sample_data` slot.
}
\description{
Calculate normalization factors from a phyloseq object to scale the raw
library sizes. Inherited from metagenomeSeq `calcNormFactors` function
which performs the Cumulative Sum Scaling normalization.
}
\seealso{
[metagenomeSeq::calcNormFactors()] for details.
}
