% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_DESeq2}
\alias{norm_DESeq2}
\title{norm_DESeq2}
\usage{
norm_DESeq2(object, method = c("ratio", "poscounts", "iterate"), ...)
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}

\item{method}{Method for estimation: either "ratio", "poscounts", or
"iterate". "ratio" uses the standard median ratio method introduced in DESeq.
The size factor is the median ratio of the sample over a "pseudosample": for
each gene, the geometric mean of all samples. "poscounts" and "iterate" offer
alternative estimators, which can be used even when all features contain a
sample with a zero (a problem for the default method, as the geometric mean
becomes zero, and the ratio undefined). The "poscounts" estimator deals with
a feature with some zeros, by calculating a modified geometric mean by taking
the n-th root of the product of the non-zero counts. This evolved out of use
cases with Paul McMurdie's phyloseq package for metagenomic samples. The
"iterate" estimator iterates between estimating the dispersion with a design
of ~1, and finding a size factor vector by numerically optimizing the
likelihood of the ~1 model.}

\item{...}{other parameters for DESeq2
\code{\link[DESeq2]{estimateSizeFactors}} function.}
}
\value{
A new column containing the chosen DESeq2-based normalization factors
is added to the phyloseq `sample_data` slot.
}
\description{
Calculate normalization factors from a phyloseq object to scale the raw
library sizes. Inherited from DESeq2 `estimateSizeFactors` function.
}
\examples{
\dontrun{
data(ps_stool_16S)
# Calculate the normalization factors
ps_stool_16S <- norm_DESeq2(object = ps_stool_16S, method = "poscounts")

# The phyloseq object now contains the normalization factors:
normFacts <- sample_data(ps_stool_16S)[,"NF.poscounts"]
head(normFacts)

# VERY IMPORTANT: to convert normalization factors to scaling factors divide
# them by the library sizes and renormalize.
scaleFacts = normFacts / sample_sums(ps_stool_16S)
# Renormalize: multiply to 1
scaleFacts = scaleFacts/exp(mean(log(scaleFacts)))
}
}
\seealso{
\code{\link[DESeq2]{estimateSizeFactors}} for details.
}
