% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_edgeR}
\alias{norm_edgeR}
\title{norm_edgeR}
\usage{
norm_edgeR(
  object,
  method = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none"),
  refColumn = NULL,
  logratioTrim = 0.3,
  sumTrim = 0.05,
  doWeighting = TRUE,
  Acutoff = -1e+10,
  p = 0.75,
  ...
)
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}

\item{method}{normalization method to be used. Choose between "TMM",
"TMMwsp", "RLE", "upperquartile", "posupperquartile" or "none".}

\item{refColumn}{column to use as reference for \code{method="TMM"}. Can be a column number or a numeric vector of length \code{nrow(object)}.}

\item{logratioTrim}{amount of trim to use on log-ratios ("M" values) for \code{method="TMM"}}

\item{sumTrim}{amount of trim to use on the combined absolute levels ("A" values) for \code{method="TMM"}}

\item{doWeighting}{logical, whether to compute (asymptotic binomial precision) weights for \code{method="TMM"}}

\item{Acutoff}{cutoff on "A" values to use before trimming for \code{method="TMM"}}

\item{p}{percentile (between 0 and 1) of the counts that is aligned when \code{method="upperquartile"}}

\item{...}{other arguments that are not currently used.}
}
\value{
A new column containing the chosen edgeR-based normalization factors
is added to the phyloseq `sample_data` slot. The effective library sizes for
use in downstream analysis must be multiplied by the normalization factors.
}
\description{
Calculate normalization factors from a phyloseq object to scale the raw
library sizes. Inherited from edgeR `calcNormFactors` function.
}
\seealso{
[edgeR::calcNormFactors()] for details.
}
