% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_edgeR}
\alias{norm_edgeR}
\title{norm_edgeR}
\usage{
norm_edgeR(
  object,
  method = c("TMM", "TMMwsp", "RLE", "upperquartile", "posupperquartile", "none"),
  refColumn = NULL,
  logratioTrim = 0.3,
  sumTrim = 0.05,
  doWeighting = TRUE,
  Acutoff = -1e+10,
  p = 0.75,
  ...
)
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}

\item{method}{normalization method to be used. Choose between "TMM",
"TMMwsp", "RLE", "upperquartile", "posupperquartile" or "none".}

\item{refColumn}{column to use as reference for \code{method="TMM"}. Can be a column number or a numeric vector of length \code{nrow(object)}.}

\item{logratioTrim}{the fraction (0 to 0.5) of observations to be trimmed from each tail of the distribution of log-ratios (M-values) before computing the mean. Used by \code{method="TMM"} for each pair of samples.}

\item{sumTrim}{the fraction (0 to 0.5) of observations to be trimmed from each tail of the distribution of A-values before computing the mean. Used by \code{method="TMM"} for each pair of samples.}

\item{doWeighting}{logical, whether to use (asymptotic binomial precision) weights when computing the mean M-values. Used by \code{method="TMM"} for each pair of samples.}

\item{Acutoff}{minimum cutoff applied to A-values. Count pairs with lower A-values are ignored. Used by \code{method="TMM"} for each pair of samples.}

\item{p}{numeric value between 0 and 1 specifying which quantile of the counts should be used by \code{method="upperquartile"}.}

\item{...}{other arguments are not currently used.}
}
\value{
A new column containing the chosen edgeR-based normalization factors
is added to the phyloseq `sample_data` slot. The effective library sizes for
use in downstream analysis must be multiplied by the normalization factors.
}
\description{
Calculate normalization factors from a phyloseq object to scale the raw
library sizes. Inherited from edgeR `calcNormFactors` function.
}
\seealso{
[edgeR::calcNormFactors()] for details.
}
