% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm_TSS}
\alias{norm_TSS}
\title{norm_TSS}
\usage{
norm_TSS(object)
}
\arguments{
\item{object}{phyloseq object containing the counts to be normalized.}
}
\value{
A new column containing the TSS normalization factors is added to the
phyloseq `sample_data` slot.
}
\description{
Calculate the raw library sizes from a phyloseq object. If used to divide
counts, known as Total Sum Scaling normalization (TSS).
}
