% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{plotMD}
\alias{plotMD}
\title{plotMD}
\usage{
plotMD(data, difference = NULL, split = TRUE)
}
\arguments{
\item{data}{the `list` output of the \code{fitModels} function or a
`data.frame` object with 'Model', 'Y', 'Y0', 'MD', and 'ZPD'
columns containing the model name, the observed values for the mean and the
zero proportion and the differences between observed and estimated values.}

\item{difference}{character vector, either 'MD' or 'ZPD' to plot the
differences between estimated and observed mean counts or the differences
between estimated zero probability and observed zero proportion.}

\item{split}{Display each model mean differences in different facets (default
= TRUE). If FALSE, points are not displayed for a more clear representation.}
}
\value{
a `ggplot` object.
}
\description{
A function to plot mean difference (MD) and zero probability difference (ZPD)
values between estimated and observed values.
}
\examples{
\donttest{
data(ps_stool_16S)

# Estimate the counts assuming several distributions
GOF_stool_16S <- fitModels(object = ps_stool_16S, models = c("NB","ZINB",
"DM","ZIG","HURDLE"), scale_ZIG = c("median","default"), scale_HURDLE =
c("median","default"))

# Plot the results
plotMD(data = GOF_stool_16S, difference = "MD", split = TRUE)
plotMD(data = GOF_stool_16S, difference = "ZPD", split = TRUE)
}
}
\seealso{
\code{\link{fitModels}} and \code{\link{RMSE}} for the model
estimations and the RMSE computations respectively. \code{\link{plotRMSE}}
for the graphical evaluation of the RMSE values.
}
