% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistribution.R
\name{prepareObserved}
\alias{prepareObserved}
\title{prepareObserved}
\usage{
prepareObserved(counts, scale = NULL)
}
\arguments{
\item{counts}{a matrix of counts with features (OTUs, ASVs, genes) by row and
samples by column.}

\item{scale}{If specified it refers to the character vector used in
`fitHURDLE` function. Either 'median' or 'default' to choose between the
median library size or one million as scaling factors for raw counts.}
}
\value{
A data frame containing the continuity corrected logarithm for the
raw count mean values for each taxon of the `counts` matrix in the `Y` column
and the observed zero rate in the `Y0` column. If `scale` is specified the
continuity corrected logarithm for the mean CPM (`scale = "default"`) or the
mean counts per median library size (`scale = "median"`) is computed instead.
}
\description{
Average count value continuity corrected logarithms and fraction of zeroes by
taxon.
}
\examples{
# Let's generate some random count data
nlibs <- 10
ntaxa <- 100
lambda = 2
counts <- matrix(rpois(ntaxa*nlibs, lambda),
                 nrow = ntaxa,
                 ncol = nlibs)

# observed values
o_values <- prepareObserved(counts)
head(o_values)
}
