% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_stool_16S.R
\docType{data}
\name{ps_stool_16S}
\alias{ps_stool_16S}
\title{(Data) 33 Stool samples of 16S rRNA (HMP 2012)}
\format{
An object of class \code{"phyloseq"}
}
\usage{
data(ps_stool_16S)
}
\description{
A demonstrative purpose dataset containing microbial abundances for a total
of 77 OTUs. The 32 Stool samples belong to the Human Microbiome Project.
This particular subset contains the SEX = "Male", RUN_CENTER = "BI",
and VISITNO = "1" samples. It is possible to obtain the same dataset after
several filters using the [HMP16Data] Bioconductor package.
}
\keyword{datasets}
