% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(differences)
}
\arguments{
\item{differences}{a vector of differences.}
}
\value{
RMSE value
}
\description{
Computes the Root Mean Square Error (RMSE) from a vector of differences.
}
\examples{
# Generate the data.frame of Mean Differences and Zero Probability Difference
MD_df <- data.frame(MD = rpois(10, 5), ZPD = runif(10, -1, 1))

# Calculate RMSE for MD and ZPD values
RMSE(MD_df[, "MD"])
RMSE(MD_df[, "ZPD"])
}
\seealso{
\code{\link{prepareObserved}} and \code{\link{meanDifferences}}.
}
