% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_ALDEx2.R
\name{set_ALDEx2}
\alias{set_ALDEx2}
\title{set_ALDEx2}
\usage{
set_ALDEx2(
  pseudo_count = FALSE,
  conditions = NULL,
  mc.samples = 128,
  test = "t",
  denom = "iqlr",
  norm = "TSS",
  expand = TRUE
)
}
\arguments{
\item{pseudo_count}{add 1 to all counts if TRUE (default
\code{pseudo_count = FALSE}).}

\item{conditions}{A character vector. A description of the data structure used
for testing. Typically, a vector of group labels. For \code{aldex.glm}, use
a \code{model.matrix}.}

\item{mc.samples}{An integer. The number of Monte Carlo samples to use when
estimating the underlying distributions. Since we are estimating central
tendencies, 128 is usually sufficient.}

\item{test}{A character string. Indicates which tests to perform. "t" runs
Welch's t test while "wilcox" runs Wilcoxon test.}

\item{denom}{A character string. Indicates which features to retain as the
denominator for the Geometric Mean calculation. Using "iqlr" accounts for
data with systematic variation and centers the features on the set features
that have variance that is between the lower and upper quartile of variance.
Using "zero" is a more extreme case where there are many non-zero features in
one condition but many zeros in another. In this case the geometric mean of
each group is calculated using the set of per-group non-zero features.}

\item{norm}{name of the normalization method used to compute the
normalization factors to use in the differential abundance analysis. If
\code{norm} is equal to "TMM", "TMMwsp", "RLE", "upperquartile",
"posupperquartile", "CSSmedian", "CSSdefault", "TSS" the scaling factors are
automatically transformed into normalization factors.}

\item{expand}{logical, if TRUE create all combinations of input parameters
(default \code{expand = TRUE})}
}
\value{
A named list containing the set of parameters for
\code{DA_ALDEx2} method.
}
\description{
Set the parameters for ALDEx2 differential abundance detection method.
}
\examples{
# Set some basic combinations of parameters for ALDEx2
base_ALDEx2 <- set_ALDEx2(conditions = "group")
# Set a specific set of normalization for ALDEx2 (even of other
# packages!)
setNorm_ALDEx2 <- set_ALDEx2(conditions = "group", norm = c("TSS", "TMM"))
# Set many possible combinations of parameters for ALDEx2
all_ALDEx2 <- set_ALDEx2(conditions = "group", denom = c("iqlr", "zero"),
    test = c("t", "wilcox"))
}
\seealso{
\code{\link{DA_ALDEx2}}
}
