% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_metagenomeSeq.R
\name{set_metagenomeSeq}
\alias{set_metagenomeSeq}
\title{set_metagenomeSeq}
\usage{
set_metagenomeSeq(
  pseudo_count = FALSE,
  design = NULL,
  coef = 2,
  norm = c("CSSmedian", "CSSdefault"),
  expand = TRUE
)
}
\arguments{
\item{pseudo_count}{add 1 to all counts if TRUE (default
\code{pseudo_count = FALSE}).}

\item{design}{The model for the count distribution. Can be the variable name,
or a character similar to "~ 1 + group", or a formula, or a `model.matrix`
object.}

\item{coef}{integer or character index vector indicating which coefficients
of the linear model are to be tested equal to zero.}

\item{norm}{name of the normalization method used to compute the
scaling factors to use in the differential abundance analysis. If \code{norm}
is equal to "ratio", "poscounts", or "iterate" the normalization factors are
automatically transformed into scaling factors.}

\item{expand}{logical, if TRUE create all combinations of input parameters
(default \code{expand = TRUE})}
}
\value{
A named list containing the set of parameters for
\code{DA_metagenomeSeq} method.
}
\description{
Set the parameters for metagenomeSeq differential abundance detection method.
}
\examples{
# Set some basic combinations of parameters for metagenomeSeq
base_mgs <- set_metagenomeSeq(design = ~ group, coef = 2)
# Set a specific set of normalization for metagenomeSeq (even of other
# packages!)
setNorm_mgs <- set_metagenomeSeq(design = ~ group, coef = 2,
    norm = c("CSSmedian", "TMM"))
# Set many possible combinations of parameters for metagenomeSeq
all_mgs <- set_metagenomeSeq(pseudo_count = c(TRUE, FALSE), design = ~ group,
    coef = 2, norm = c("CSSmedian", "CSSdefault", "TMM", "TSS"))
}
\seealso{
\code{\link{DA_metagenomeSeq}}
}
